Refactor packages and add app icon
