+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
             android:label="@string/app_name"
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
-            android:name=".application.ui.MainApplication"
+            android:name=".application.MainApplication"
             android:theme="@style/AppTheme">
         <activity android:name=".application.ui.MainActivity">
             <intent-filter>
