+++ app/src/main/java/com/igmob/android/marvelapi/application/MainApplication.kt
@@ -1,8 +1,8 @@
-package com.igmob.android.marvelapi.application.ui
+package com.igmob.android.marvelapi.application
 
 import android.app.Application
 import com.facebook.drawee.backends.pipeline.Fresco
-import com.igmob.android.marvelapi.infrastructure.networking.NetworkingModule
+import com.igmob.android.marvelapi.infrastructure.core.networking.NetworkingModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
