+++ app/src/main/java/com/igmob/android/marvelapi/domain/characterslist/CharactersListRepository.kt
@@ -1,0 +1,9 @@
+package com.igmob.android.marvelapi.domain.characterslist
+
+import com.igmob.android.marvelapi.domain.core.Character
+import com.igmob.android.marvelapi.domain.core.Resource
+
+interface CharactersListRepository {
+    suspend fun getCharacters(): Resource<List<Character>>
+}
+
