--- app/src/main/java/com/igmob/android/marvelapi/domain/core/CharactersListRepository.kt
@@ -1,6 +1,0 @@
-package com.igmob.android.marvelapi.domain.core
-
-interface CharactersListRepository {
-    suspend fun getCharacters(): Resource<List<Character>>
-}
-
