+++ app/src/main/java/com/igmob/android/marvelapi/domain/usecases/GetCharactersListUseCase.kt
@@ -1,6 +1,6 @@
 package com.igmob.android.marvelapi.domain.usecases
 
-import com.igmob.android.marvelapi.domain.core.CharactersListRepository
+import com.igmob.android.marvelapi.domain.characterslist.CharactersListRepository
 
 class GetCharactersListUseCase(
     private val repository: CharactersListRepository
