+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Marvelapi</string>
+    <string name="app_name">Marvel Api</string>
     <string name="view_more">View more</string>
 </resources>
