+++ app/src/test/java/com/igmob/android/marvelapi/infrastructure/TestUtils.kt
@@ -1,4 +1,4 @@
-package com.igmob.android.marvelapi.infrastructure.networking
+package com.igmob.android.marvelapi.infrastructure
 
 import com.squareup.moshi.Moshi
